Inline Def into Expr, move parts to expr module
