--- vunk-parser/src/ast/decl.rs
@@ -1,17 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum DeclType {
-    TypeName(String),
-    Func { args: Vec<DeclArg>, retty: String },
-}
-
-#[derive(Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct DeclArg {
-    pub name: Option<String>,
-    pub ty: DeclType,
-}
