+++ vunk-parser/src/ast/expr.rs
@@ -2,8 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::ast::decl::DeclType;
-use crate::ast::def::Def;
 use crate::ast::generic::WhereClause;
 use crate::ast::op::BinaryOp;
 use crate::ast::op::UnaryOp;
@@ -63,5 +61,30 @@ pub enum Expr {
         whereclause: Option<WhereClause>,
     },
 
-    Def(Def),
+    Def {
+        lhs: String,
+        rhs: DefRhs,
+    },
 }
+
+#[derive(Debug)]
+#[cfg_attr(test, derive(PartialEq))]
+pub enum DeclType {
+    TypeName(String),
+    Func { args: Vec<DeclArg>, retty: String },
+}
+
+#[derive(Debug)]
+#[cfg_attr(test, derive(PartialEq))]
+pub struct DeclArg {
+    pub name: Option<String>,
+    pub ty: DeclType,
+}
+
+#[derive(Debug)]
+#[cfg_attr(test, derive(PartialEq))]
+pub struct DefRhs {
+    pub args: Vec<(String, DeclType)>,
+    pub expr: Box<Expr>,
+}
+
