+++ vunk-parser/src/ast/mod.rs
@@ -2,8 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-pub mod decl;
-pub mod def;
 pub mod expr;
 pub mod generic;
 pub mod literal;
